Implements project pagination with caching
Improves pagination for the project list for improved performance,
especially with large numbers of projects.

Improves UX by loading projects in smaller chunks.

The Speckle service is updated to cache projects based on both
the user token and the cursor, to avoid duplicate calls.
(first implementation was on cursor only which had me scratching
 my head for while......til i remembered)

Adds logging for debugging cursor issues.